==> 1-Process-Historic-AADT.ipynb <==
# import pandas
import pandas as pd

# Pull data from AADT excel spreadsheet and put into dataframes

# GLOBAL VARIABLES

# excel spreadsheet copied to data/udot folder from https://drive.google.com/file/d/1rDXm0ObugGR1zXgWUuVbzWHNt-Xs1xru/view
fnExcelAADTHistory = 'data/AADTHistory.xlsx'

# create connection to Excel file and get list of sheets
xl_file1 = pd.ExcelFile(fnExcelAADTHistory)

# display sheet names
print('Excel sheet names: ' + str(xl_file1.sheet_names))

Note: AADT for 2021 is in a separate sheet called 'AADT_2021'. AADT for years 1981-2020 are in AADT_Rounded and AADT_Unrounded. AADT unrounded is what we will use, since we want to forecast from unrounded values

# put data in to series of dataframes (dfsExcelSheets)
dfsExcelSheets = {sheet_name: xl_file1.parse(sheet_name) for sheet_name in xl_file1.sheet_names}

# Get AADT for years prior to 2021

# use temp dataframe _df as a copy of excel sheet
_df = dfsExcelSheets['AADT_Unrounded'].copy()

# get all columns that start with 'AADT'
aadt_cols = [col for col in _df.columns if col.startswith('AADT')]

# define the id columns that we want to keep
id_vars = ['Route', 'Beg MP', 'End MP']

# melt the DataFrame - melt is a reverse pivot
dfUnder2021 = _df.melt(id_vars=id_vars, value_vars=aadt_cols, var_name='YEAR', value_name='AADT')

display (dfUnder2021)

# replace 'AADT' in the 'YEAR' column and convert to integer
dfUnder2021['YEAR'] = dfUnder2021['YEAR'].str.replace('AADT', '').astype(int)

# rename columns to match 2021 data
dfUnder2021.rename(columns={'Route':'ROUTE_ID','Beg MP':'FROM_MEASURE','End MP':'TO_MEASURE'}, inplace=True)

display(dfUnder2021)

# Process 2021 AADTs

AADTs for 2021 are different from other years and are in their own sheet called 'AADT 2021'

# show snippet of 2021 sheet
dfsExcelSheets['AADT_2021']

# PREPARE 2021 DATA
df2021 = dfsExcelSheets['AADT_2021'].copy(0)

# add column for YEAR
df2021['YEAR'] = 2021

# rename AADT_2021 to AADT to match data from other years
df2021.rename(columns={'AADT_2021':'AADT'}, inplace=True)
df2021

# Combine under 2021 and 2021

# combine 2021 and under 2021
dfAADT = pd.concat([dfUnder2021,df2021])

#dfAADT.reset_index(inplace=True)
dfAADT.fillna('0',inplace=True)

# filter out 0 AADTs
dfAADT = dfAADT[dfAADT['AADT']>0]

# round to nearest hundredth
dfAADT['FROM_MEASURE'] = dfAADT['FROM_MEASURE'].round(2)
dfAADT['TO_MEASURE'  ] = dfAADT['TO_MEASURE'  ].round(2)

display(dfAADT)

# Filter rows with 'ROUTE_ID' containing 'PM': Positive Direction & Mainline
dfAADT_filtered = dfAADT[dfAADT['ROUTE_ID'].str.contains('PM')].copy()

# Trim 'ROUTE_ID' down to the first four characters in order to match with SEGID
dfAADT_filtered['ROUTE_ID'] = dfAADT_filtered['ROUTE_ID'].str[:4]

# sort by fields SEGID and YEAR
dfAADT_filtered = dfAADT_filtered.sort_values(by=['ROUTE_ID','FROM_MEASURE','TO_MEASURE','YEAR'])

dfAADT_filtered

# CHECK FOR DUPLICATES

# Define the columns you want to check for duplicates
columns_to_check = ['ROUTE_ID', 'FROM_MEASURE', 'TO_MEASURE', 'YEAR']

# Use the duplicated method to find duplicates in those columns
duplicates = dfAADT_filtered.duplicated(subset=columns_to_check, keep=False)

# Filter the DataFrame to only include the duplicates
dfAADT_duplicates = dfAADT_filtered[duplicates]

display(dfAADT_duplicates)

# Export

dfAADT_filtered.to_csv('intermediate/historic-aadt.csv', index=False)
